code cleaning (3) and repairs in player.json and entities in world.json because of renaming a variable
