+++ SaveFiles/player.json
@@ -5,7 +5,7 @@
         "name": "Catmon",
         "backSprite": "assets/fighters/backsprites/citizen_back.png",
         "type": "CITIZEN",
-        "battleParty": "PLAYER",
+        "fightingSide": "PLAYER",
         "ID": 0,
         "isDefeated": false,
         "maxHitpoints": 100,
@@ -19,7 +19,7 @@
         "name": "Lionmon",
         "backSprite": "assets/fighters/backsprites/undead_back.png",
         "type": "UNDEAD",
-        "battleParty": "PLAYER",
+        "fightingSide": "PLAYER",
         "ID": 0,
         "isDefeated": false,
         "maxHitpoints": 100,
