+++ SaveFiles/world.json
@@ -650,7 +650,7 @@
               "name": "Dogmon",
               "backSprite": "assets/fighters/backsprites/citizen_back.png",
               "type": "CITIZEN",
-              "battleParty": "OPPONENT",
+              "fightingSide": "OPPONENT",
               "ID": 4,
               "isDefeated": false,
               "maxHitpoints": 25,
@@ -664,7 +664,7 @@
               "name": "Bearmon",
               "backSprite": "assets/fighters/backsprites/citizen_back.png",
               "type": "CITIZEN",
-              "battleParty": "OPPONENT",
+              "fightingSide": "OPPONENT",
               "ID": 5,
               "isDefeated": false,
               "maxHitpoints": 50,
@@ -689,7 +689,7 @@
               "name": "Dogmon",
               "backSprite": "assets/fighters/backsprites/undead_back.png",
               "type": "UNDEAD",
-              "battleParty": "OPPONENT",
+              "fightingSide": "OPPONENT",
               "ID": 4,
               "isDefeated": false,
               "maxHitpoints": 25,
@@ -703,7 +703,7 @@
               "name": "Bearmon",
               "backSprite": "assets/fighters/backsprites/undead_back.png",
               "type": "UNDEAD",
-              "battleParty": "OPPONENT",
+              "fightingSide": "OPPONENT",
               "ID": 5,
               "isDefeated": false,
               "maxHitpoints": 50,
