+++ src/Entity/Items/SupportItem.java
@@ -1,5 +1,4 @@
 package Entity.Items;
-import Entity.Item;
 
 public abstract class SupportItem extends Item {
     private final boolean offensive = false;
