+++ src/Entity/Items/ThrowableItem.java
@@ -1,5 +1,4 @@
 package Entity.Items;
-import Entity.Item;
 
 public abstract class ThrowableItem extends Item {
     private final boolean offensive = true;
