+++ src/Frames/BasicPanel.java
@@ -4,9 +4,9 @@ import BattleSystem.BattleSystem;
 import BattleSystem.Fighter;
 import BattleSystem.Fighters.Exorcist;
 import BattleSystem.FightingSide;
+import Entity.Entities.OpponentEntity;
+import Entity.Entities.PlayerEntity;
 import Entity.FighterInventory;
-import Entity.OpponentEntity;
-import Entity.PlayerEntity;
 import Frames.BattleUI.BattlePane;
 import Frames.WorldUI.WorldPane;
 import Observer.Observer;
