+++ src/Frames/WorldUI/WorldPane.java
@@ -5,9 +5,9 @@ import BattleSystem.Fighters.Citizen;
 import BattleSystem.Fighters.Exorcist;
 import BattleSystem.Fighters.Undead;
 import BattleSystem.FightingSide;
-import Entity.InteractionType;
-import Entity.OpponentEntity;
-import Entity.PlayerEntity;
+import Entity.Entities.InteractionType;
+import Entity.Entities.OpponentEntity;
+import Entity.Entities.PlayerEntity;
 import Frames.TextBox.TextBox;
 import Observer.ObserveType;
 import Observer.Observer;
