+++ src/ReadAndWrite/PlayerOperations/ReadPlayerFromJson.java
@@ -4,9 +4,9 @@ import BattleSystem.Fighter;
 import BattleSystem.Fighters.Citizen;
 import BattleSystem.Fighters.Exorcist;
 import BattleSystem.Fighters.Undead;
+import Entity.Entities.PlayerEntity;
 import Entity.FighterInventory;
 import Entity.Inventory;
-import Entity.PlayerEntity;
 import ReadAndWrite.ReadObjectFromFile;
 import Worlds.Coordinates;
 import com.google.gson.Gson;
