+++ src/ReadAndWrite/PlayerOperations/WritePlayerToJson.java
@@ -1,8 +1,9 @@
 package ReadAndWrite.PlayerOperations;
 
-import Entity.PlayerEntity;
+import Entity.Entities.PlayerEntity;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+
 import java.io.FileWriter;
 import java.io.IOException;
 public class WritePlayerToJson {
